requirements.txt
aiofiles==24.1.0
aiolimiter==1.1.0
blessed==1.20.0
colorama==0.4.6
editor==1.6.6
inquirer==3.3.0
pyaes==1.6.1
pyasn1==0.6.0
PyYAML==6.0.2
readchar==4.1.0
rsa==4.9
runs==1.2.2
six==1.16.0
Telethon==1.36.0
tqdm==4.66.5
wcwidth==0.2.13
xmod==1.8.1
